refactor(editor): shift to blocks editor
